x-8 - maybe fix amd64 geolocation bug?

+++ modules/x-8/x-8.cpp
@@ -315,7 +315,7 @@ void X8Dialogue::locationSuccess(GeoLocationResult *result)
 	reply += "\n";
 
 	reply += "\t Context: ";
-	sprintf(ctx,"%x",(uint32_t)result->getObject());
+	sprintf(ctx,"%x",(uint32_t)((intptr_t)result->getObject()));
 	reply += ctx;
 	reply += "\n";
